Add files via upload
We make 3 functions for FIFO, LRU, and Optimal each function conation the algorithms and conditions of each method. In the main, we call the functions in the switch case and we used a 2D array to print the final shape. Rogina made the FIFO algorithm and I made the Optimal algorithm and we work together in the main and LRU.
Salma Osama 2019/05060
Rogina Michelle 2019/02401